+++ src/Campaigns/resources/admin/components/CampaignDetailsPage/Tabs/Forms.tsx
@@ -1,8 +1,10 @@
 import DonationFormsListTable from '../../../../../../DonationForms/V2/resources/components/DonationFormsListTable';
+import {useCampaignEntityRecord} from '@givewp/campaigns/utils';
 
 /**
  * @unreleased
  */
 export default () => {
-    return <DonationFormsListTable />;
+    const entity = useCampaignEntityRecord();
+    return <DonationFormsListTable entity={entity} />;
 };
