+++ src/Campaigns/resources/utils.ts
@@ -1,0 +1,23 @@
+import {useEntityRecord} from '@wordpress/core-data';
+import {Campaign} from '@givewp/campaigns/admin/components/types';
+
+/**
+ * @unreleased
+ */
+export function useCampaignEntityRecord(campaignId?: number) {
+    const urlParams = new URLSearchParams(window.location.search);
+
+    const {
+        record: campaign,
+        hasResolved,
+        save,
+        edit,
+    }: {
+        record: Campaign;
+        hasResolved: boolean;
+        save: () => any;
+        edit: (data: Campaign) => void;
+    } = useEntityRecord('givewp', 'campaign', campaignId ?? urlParams.get('id'));
+
+    return {campaign, hasResolved, save, edit};
+}
